Generate a relative abundance matrix with sample meta data and OTU abundances from a phyloseq object.

The R package phyloseq streamlines the storage and analysis of microbiome sequence data. This function takes a phyloseq object and extracts the OTU table and the sample meta data and combines them into one relative abundance matrix with rows corresponding to samples, meta data on the left-hand side, and OTU relative abundances on the right-hand side.

Usage

relab_phyloseq(phyloseq_object)

Arguments

phyloseq_object

A phyloseq object containing both an OTU table (otu_table) and sample metadata (sample_data).

Value

A data frame with rows representing samples and columns representing sample data categories or OTU relative abundances.

Examples

otu_table = relab_phyloseq(FAVA::xue_phyloseq)
otu_table[1:10, 1:6]
#>      subject timepoint Actinomyces_sp_58647 Actinomyces_sp_59161
#> sa1      XBA         1         2.995750e-05         1.945213e-05
#> sa2      XBA         8         0.000000e+00         0.000000e+00
#> sa3      XBA        15         0.000000e+00         0.000000e+00
#> sa4      XBA        22         0.000000e+00         0.000000e+00
#> sa5      XBA        23         0.000000e+00         0.000000e+00
#> sa6      XBA        24         4.298487e-05         0.000000e+00
#> sa7      XBA        25         0.000000e+00         0.000000e+00
#> sa8      XBA        26         0.000000e+00         0.000000e+00
#> sa9      XBA        27         0.000000e+00         0.000000e+00
#> sa10     XBA        28         0.000000e+00         0.000000e+00
#>      Actinomyces_sp_59944 Aggregatibacter_aphrophilus_58143
#> sa1          4.023978e-05                      9.317754e-05
#> sa2          0.000000e+00                      0.000000e+00
#> sa3          0.000000e+00                      0.000000e+00
#> sa4          0.000000e+00                      0.000000e+00
#> sa5          0.000000e+00                      0.000000e+00
#> sa6          0.000000e+00                      0.000000e+00
#> sa7          0.000000e+00                      5.134258e-05
#> sa8          5.551497e-05                      0.000000e+00
#> sa9          0.000000e+00                      0.000000e+00
#> sa10         0.000000e+00                      0.000000e+00